Move protocol formatting to platform boundary
